Backend: FirmController, LocationController, DeviceController, custom jwt token for device implementation, service method for update location by jwt token and epc code

+++ backend/src/main/java/kcz/rfid/backend/controller/DeviceController.java
@@ -1,6 +1,9 @@
 package kcz.rfid.backend.controller;
 
+import kcz.rfid.backend.config.security.DeviceAuthenticationToken;
+import kcz.rfid.backend.exception.ResourceNotFoundException;
 import kcz.rfid.backend.model.dto.DeviceDto;
+import kcz.rfid.backend.model.dto.DeviceLocationDto;
 import kcz.rfid.backend.model.dto.RegisterDeviceDto;
 import kcz.rfid.backend.model.entity.DeviceEntity;
 import kcz.rfid.backend.model.entity.UserEntity;
@@ -15,6 +18,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
@@ -70,7 +74,7 @@ public class DeviceController {
     public ResponseEntity<DeviceDto> getDeviceById(@PathVariable UUID id, @AuthenticationPrincipal UserDetails userDetails) {
         DeviceEntity device = deviceService.findById(id);
         if (device == null) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            throw new ResourceNotFoundException("Device with id " + id + " not found");
         }
         if (!userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ROOT"))) {
             return new ResponseEntity<>(deviceMapper.mapToDto(device), HttpStatus.OK);
@@ -89,7 +93,11 @@ public class DeviceController {
     public ResponseEntity<DeviceDto> updateDevice(@RequestBody String deviceName, @AuthenticationPrincipal UserDetails userDetails, @PathVariable UUID id) {
         if (userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
             UserEntity admin = userService.findUserByUsername(userDetails.getUsername());
+
             DeviceEntity device = deviceService.findById(id);
+            if(device == null) {
+                throw new ResourceNotFoundException("Device with id " + id + " not found");
+            }
 
             if(!device.getFirm().equals(admin.getFirm())) {
                 return new ResponseEntity<>(HttpStatus.FORBIDDEN);
@@ -106,12 +114,35 @@ public class DeviceController {
     public ResponseEntity<DeviceDto> deleteDevice(@PathVariable UUID id, @AuthenticationPrincipal UserDetails userDetails) {
         if (userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
             UserEntity admin = userService.findUserByUsername(userDetails.getUsername());
+
             DeviceEntity device = deviceService.findById(id);
+            if(device == null) {
+                throw new ResourceNotFoundException("Device with id " + id + " not found");
+            }
 
             if(!device.getFirm().equals(admin.getFirm())) {
                 return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             }
             deviceService.delete(device);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
+
+    @PutMapping("/updateLocation")
+    public ResponseEntity<DeviceDto> updateDeviceLocation(@RequestBody DeviceLocationDto dto) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        if (auth instanceof DeviceAuthenticationToken deviceAuth) {
+            UUID deviceId = deviceAuth.getDeviceId();
+            UUID firmId = deviceAuth.getFirmId();
+
+            if (dto.getId().equals(deviceId) && dto.getFirmId().equals(firmId)) {
+                firmService.updateDeviceLocation(dto);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            else {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }
         return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
